Added keyboard combination to send newline to Tmux
Some interpreters/REPLs require additional newlines after definitions in code.
In Python there is often code where the interpreter cannot deduce if the input
is complete:

def foo( x ):
	print( "yes" ) #<- send this chunk to tmux

	print( "no" ) #<- but don't forget this part of the function

This function needs to be sent to the Python interpreter in two pieces because
there's an empty line inbetween the two print statements (unless we use visual
mode).

In order to help with these cases the module now defines the key combination
control+C+Enter to send a newline to the interpreter.